chore(install): move various interfaces closer to their usage
