+++ internal/install/interfaces.go
@@ -2,6 +2,7 @@ package install
 
 import (
 	"context"
+	"net/url"
 
 	"github.com/newrelic/newrelic-cli/internal/install/types"
 )
@@ -10,8 +11,33 @@ type ConfigValidator interface {
 	Validate(ctx context.Context) error
 }
 
+// Discoverer is reesponsible for discovering informataion about the host system.
+type Discoverer interface {
+	Discover(context.Context) (*types.DiscoveryManifest, error)
+}
+
+// FileFilterer determines the existence of files on the underlying filesystem.
+type FileFilterer interface {
+	Filter(context.Context, []types.OpenInstallationRecipe) ([]types.OpenInstallationLogMatch, error)
+}
+
+type Prompter interface {
+	PromptYesNo(msg string) (bool, error)
+	MultiSelect(msg string, options []string) ([]string, error)
+}
+
+type RecipeFileFetcher interface {
+	FetchRecipeFile(recipeURL *url.URL) (*types.OpenInstallationRecipe, error)
+	LoadRecipeFile(filename string) (*types.OpenInstallationRecipe, error)
+}
+
 type RecipeRecommender interface {
 	Recommend(ctx context.Context, m *types.DiscoveryManifest) ([]types.OpenInstallationRecipe, error)
+}
+
+// RecipeValidator validates installation of a recipe.
+type RecipeValidator interface {
+	ValidateRecipe(context.Context, types.DiscoveryManifest, types.OpenInstallationRecipe) (entityGUID string, err error)
 }
 
 type RecipeVarPreparer interface {
