--- internal/install/ux/prompter.go
@@ -1,6 +1,0 @@
-package ux
-
-type Prompter interface {
-	PromptYesNo(msg string) (bool, error)
-	MultiSelect(msg string, options []string) ([]string, error)
-}
